+++ src/LoRa_AT_Common.h
@@ -9,7 +9,7 @@
 #define SRC_LORA_AT_COMMON_H_
 
 // The current library version number
-#define LORA_AT_VERSION "0.4.2"
+#define LORA_AT_VERSION "0.4.3"
 
 #if defined(SPARK) || defined(PARTICLE)
 #include "Particle.h"